first commit - PoC
This commit contains a proof of concept. I thought this would be a cool
idea, so I looked into a few things and am now set on publishing it.

Found a nice demo for Windows Phone showing how to simulate snow fall,
brought that over the Windows 8 and added some libraries and it works
the same. I'm happy about that and must remember to reference the demo I
found - I believe it was a Windows Phone Geek tutorial.

I determined that I'm able to get today's date, subtract it from date of
christmas and figure out how many days and hours are left until
Christmas.

TODO:
Format the time span difference as Days, Hours, Minutes and Seconds
until Christmas.

Add references and credits.

Artwork