Added snap view, removed ads
This is the free version with no ads and it now features snap view. I'm
going to look into live tile now and if it seems like I can implement it
in a couple of hours I'll do it today and publish a nice update early
this week. Otherwise I'll probably publish this update with snap view
and tell people about the ad supported version which is being certified
right now. I've had the app published for less than 2 weeks now, 55
cumulative downloads and Windows 8 was just released yesterday. The
Surface sold very well and the number of downloads has increased greatly
for all apps in the store. I'm trying to monetize some Windows 8 apps so
I want ot have a good looking free version and the "donate" version will
be ad supported.