Added live tile.
Discovered that if you keep the app running in sleep mode you can use a
timer in your code to periodically toggle the tile (e.g., update,
clear). However, if your app is completely shut down you have to invoke
a background task. I started working on that but haven't yet tested it.
The minimum amount of time in between the background interrupts is 15
minutes, which is kind of weak. Might have to look into grabbing the
time from some URL, grabbing the date from it, computing time until
Christmas and displaying that. 15 minutes is kind of long, maybe
acceptable though.