live tiles updating every minute
The format of the tiles is incomplete though.

I ended up using a background task, like I originally thought, which
will fire every 15 minutes but each time it fires I schedule a bunch of
notifications to cycle through on the live tile.

To do this I'm creating an XML document from a string composed of the
time remaining until Christmas and today's date and then binding that
XML document to the tile's attributes. My text is wrapping right now,
which I think is an issue with the tile template I'm using but I haven't
landed on one that I like yet.

Also, I included a new library called Clock.WinRT which I got from the
link below and is doing alot of the tile work for me. It is also where I
deal with formatting the tile but since it is a separate library, each
time I update my code I have to build the library, dereference it from
my Christmas Countdown project, and then reference the newer build. It's
a somewhat tedious process for which I'd like to find a quicker way.