Added logic to work around Christmas day
Added logic in OnNavigatedTo handler in MainPage.cs to handle a couple
of cases:
(1) On Christmas day, start the app out with a Merry Christmas message.
(2) On 12/26 - 12/31 adjust the countdown accordingly so that we are
counting down until Christmas of next year, not the one that occured
just a few days ago.

TODO:
- Add logic to adjust UI and display a Merry Christmas message on
Christmas day. Shouldn't take long at all, and it would be good to
incorporate snap view.
- Design and use a live tile that will display a Merry Christmas message
on Christmas day.