Fixed about page, started settings page
I think this is actually quite a big update but I didn't document
everything I did.

The about page was out of whack and looked ugly.
I found a bug such that the live tile was not correctly calculating for
the days after Christmas but before the new year. I fixed the bug, but
unfortunately I was too late for this year.

In the settings page I'm going to include:
- Change background color (only between 6 colors)
- Change live tile style (only between a few styles)

Future updates I'd like to do include:
- Gift list functionality