Fixed live tile bug, worked on about page
I fixed the live tile bug and pushed an update to the store with a fix.
Hopefully the bug is fixed soon.

I also worked on the about page quite a bit and am working towards
advertising the rest of our apps on the about page.

Productive night on the new Slate! These commits were done using a
Samsung Series 7 Slate running Windows 8 Pro and being operated with a
bluetooth keyboard and mouse. Great experience!